backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
httpx
langgraph
langchain-core
numpy
pydicom
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg
Pillow
qrcode
transformers
torch
google-cloud-aiplatform
google-cloud-storage
scikit-image
sqlalchemy
redis
PyJWT
psycopg2-binary
alembic
